Remove the transcribe sub-module
Whisper is the only backend, so start refactoring the backends out.
